Moves loading the bash/env file into bashrc and ensures that it is only loaded once
This fixes an issue that occurred when executing command remotely via ssh. In this case, ssh did only load the bashrc and not the bash_profile which caused the environment not to be set up correctly.

This commit makes bash_profile basically unnecessary. The file is only kept for compatibility reasons and simply sources bashrc.